refactor: 공통 모달 컴포넌트 경로 수정
